email-integration

+++ controllers/courses.js
@@ -42,12 +42,23 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access  Private
 exports.createCourse = asyncHandler(async (req, res, next) => {
   req.body.bootcamp = req.params.bootcampId; // add bootcampId to req.body
+  req.body.user = req.user.id;
+
   const bootcamp = await bootcamps.findById(req.params.bootcampId);
 
   if (!bootcamp) {
     throw new ErrorResponse(
       `No bootcamp with the id of ${req.params.bootcampId}`,
       404,
+    );
+  }
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        401,
+      ),
     );
   }
   const course = await Course.create(req.body);
@@ -64,6 +75,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 
   if (!course) {
     throw new ErrorResponse(`No course with the id of ${req.params.id}`, 404);
+  }
+
+  // Make sure user is bootcamp owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update a course ${course}`,
+        401,
+      ),
+    );
   }
 
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
@@ -84,6 +105,15 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     throw new ErrorResponse(`No course with the id of ${req.params.id}`, 404);
   }
 
+  // Make sure user is bootcamp owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete a course ${course}`,
+        401,
+      ),
+    );
+  }
   await course.deleteOne();
   res.status(200).json({ success: true, data: {} });
 });
